requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
bce-python-sdk==0.9.10
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.8.2
Django==4.1.4
django-cors-headers==4.3.1
django-simpleui==2024.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
erniebot==0.5.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
multidict==6.0.5
# mysqlclient==2.2.4
pillow==10.3.0
prompt_toolkit==3.0.47
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.8
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
tencentcloud-sdk-python==3.0.1187
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
